requirements.txt
haystack-ai
pymed
gradio
python-dotenv
transformers